Move create, rename and delete functions to library
