+++ gui/wxpython/location_wizard/wizard.py
@@ -59,7 +59,7 @@ from gui_core.wrap import SpinCtrl, SearchCtrl, StaticText, \
     TextCtrl, Button, CheckBox, StaticBox, NewId, ListCtrl, HyperlinkCtrl
 from location_wizard.dialogs import SelectTransformDialog
 
-from grass.grassdb.mapset import location_exists
+from grass.grassdb.checks import location_exists
 from grass.script import decode
 from grass.script import core as grass
 from grass.exceptions import OpenError
