+++ gui/wxpython/startup/utils.py
@@ -17,7 +17,6 @@ solve the errors etc. in a general manner).
 
 
 import os
-import shutil
 import tempfile
 import getpass
 import sys
@@ -88,61 +87,3 @@ def create_database_directory():
         pass
 
     return None
-
-
-def create_mapset(database, location, mapset):
-    """Creates a mapset in a specified location"""
-    location_path = os.path.join(database, location)
-    mapset_path = os.path.join(location_path, mapset)
-    # create an empty directory
-    os.mkdir(mapset_path)
-    # copy DEFAULT_WIND file and its permissions from PERMANENT
-    # to WIND in the new mapset
-    region_path1 = os.path.join(location_path, "PERMANENT", "DEFAULT_WIND")
-    region_path2 = os.path.join(location_path, mapset, "WIND")
-    shutil.copy(region_path1, region_path2)
-    # set permissions to u+rw,go+r (disabled; why?)
-    # os.chmod(os.path.join(database,location,mapset,'WIND'), 0644)
-
-
-def delete_mapset(database, location, mapset):
-    """Deletes a specified mapset"""
-    if mapset == "PERMANENT":
-        # TODO: translatable or not?
-        raise ValueError(
-            "Mapset PERMANENT cannot be deleted" " (whole location can be)"
-        )
-    shutil.rmtree(os.path.join(database, location, mapset))
-
-
-def delete_location(database, location):
-    """Deletes a specified location"""
-    shutil.rmtree(os.path.join(database, location))
-
-
-def rename_mapset(database, location, old_name, new_name):
-    """Rename mapset from *old_name* to *new_name*"""
-    location_path = os.path.join(database, location)
-    os.rename(
-        os.path.join(location_path, old_name), os.path.join(location_path, new_name)
-    )
-
-
-def rename_location(database, old_name, new_name):
-    """Rename location from *old_name* to *new_name*"""
-    os.rename(os.path.join(database, old_name), os.path.join(database, new_name))
-
-
-
-
-
-def get_default_mapset_name():
-    """Returns default name for mapset."""
-    try:
-        defaultName = getpass.getuser()
-        defaultName.encode("ascii")
-    except UnicodeEncodeError:
-        # raise error if not ascii (not valid mapset name)
-        defaultName = "user"
-
-    return defaultName
